Review - Split HtmlReloader in two separate Reloader classes
